+++ src/Carbunql.TypeSafe/ITableRowDefinition.cs
@@ -1,0 +1,10 @@
+﻿using Carbunql.Annotations;
+using Carbunql.Clauses;
+
+namespace Carbunql.TypeSafe;
+
+public interface ITableRowDefinition
+{
+    [IgnoreMapping]
+    TableDefinitionClause TableDefinition { get; set; }
+}
